+++ CHANGELOG.md
@@ -9,8 +9,9 @@
 * 新特性
   - 新增 ContextProvider 提供上下文参数跟踪能力
 
-* 使用方式变更
+* 规范变更
   - 由 nstater-core 的 BaseComponent 取代 AbstractComponent
+  - 不在保留全局统一的 types 目录，而有各组件自行管理类型定义
 
 * 组件升级
   - 升级 nstarter-grpc -> 0.3.x 并基于上游插件管理 grpc 依赖